🔨 refactor: 코드 분리, 타이틀 및 레이아웃 수정
